added sweetalert2 in every delete

+++ src/views/pages/company/ListCompanies.js
@@ -3,19 +3,41 @@ import CIcon from '@coreui/icons-react'
 import React, { useEffect, useState } from 'react'
 import { Link} from 'react-router-dom'
 import { toast } from 'react-toastify'
+import Swal from 'sweetalert2'
 import companyService from '../../services/company'
 
 function ListCompanies() {
   const [campanies, setCompanies] = useState([])
   const getCompanies = async () => {
     const response = await companyService.getAllCompanies();
-    setCompanies(response.data)
+    setCompanies(response.data);
   }
   const handleDelete = async (id) => {
-    const response = await companyService.removeOne(id)
-    getCompanies()
-    toast.success(response.data.message)
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger me-2'
+      },
+      buttonsStyling: false
+    })
+    
+    swalWithBootstrapButtons.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      // icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes!',
+      cancelButtonText: 'No, cancel!',
+      reverseButtons: true
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const response = await companyService.removeOne(id)
+        getCompanies()
+        toast.success(response.data.message)
+      }
+    })
   }
+  
 
   useEffect(() => {
     getCompanies()
@@ -27,9 +49,9 @@ function ListCompanies() {
       </div>
       <div className="card-body">
         <Link className='btn btn-secondary my-4' to='/company/create'>Create company</Link>
-        <table className="table table-hover">
+        <table className="table table-hover table-striped">
           <thead>
-            <tr>
+            <tr style={{backgroundColor :'#46546C', color: 'white'}}>
               <th className='col-1'>#</th>
               <th className='col-2'>Name</th>
               <th className='col-3'>E-mail</th>
@@ -43,11 +65,11 @@ function ListCompanies() {
               campanies.map((company, index) => {
                 return (
                   <tr key={index} className="">
-                    <td>{index + 1}</td>
+                    <td style={{backgroundColor :'#46546C', color: 'white'}}>{index + 1}</td>
                     <td>{company.companyName}</td>
                     <td>{company.email}</td>
                     <td>{company.role}</td>
-                    <td><img src={company.photo} width='70px' height='70px' alt='' /></td>
+                    <td><img src={company.photo} width='50px' height='50px' alt='' /></td>
                     <td>
                       <Link className='btn btn-info my-4' to={'/company/update/' + company._id}><CIcon icon={cilPen} size="lg" style={{ color: 'white' }} /></Link>
                       <button className='btn btn-danger ms-3' onClick={() => handleDelete(company._id)} ><CIcon icon={cilTrash} size="lg" style={{ color: 'white' }} /></button>
